Retrieve GitHub information.

Returns the GitHub username, stars, forks, branches, issues, labels, milestones, code of conduct, license, pull requests and releases.

get_gh_username(repo_name)

get_gh_stats(repo_name, user_name = NULL)

get_gh_branches(repo_name, user_name = NULL)

get_gh_issues(repo_name, user_name = NULL)

get_gh_labels(repo_name, user_name = NULL)

get_gh_milestones(repo_name, user_name = NULL)

get_gh_coc(repo_name, user_name = NULL)

get_gh_license(repo_name, user_name = NULL)

get_gh_pr(repo_name, user_name = NULL)

get_gh_releases(repo_name, user_name = NULL)

Arguments

repo_name
Name of the package.

user_name
GitHub user or organization name.

Examples

# get github user/organization name
get_gh_username("olsrr")

#> No encoding supplied: defaulting to UTF-8.

#> [1] "rsquaredacademy"

# get count of stars, forks & issues
get_gh_stats("dplyr", "tidyverse")

#> # A tibble: 1 x 3
#>   stars issues forks
#>   <int>  <int> <int>
#> 1  2977    115  1087

# get details of github branches
get_gh_branches("dplyr", "tidyverse")

#> # A tibble: 16 x 1
#>    branches                       
#>    <chr>                          
#>  1 3451/filter_size_mismatch_error
#>  2 3772-name_repair               
#>  3 3953_ghost_programming_vignette
#>  4 4004_dollar_gets_grouped       
#>  5 4249/bind_rows_id              
#>  6 4277/quos_vars                 
#>  7 4329-non_ascii                 
#>  8 bug-tibble-vignette            
#>  9 dev-ggplot2                    
#> 10 f-travis-4                     
#> 11 gh-pages                       
#> 12 master                         
#> 13 r-0.7.6                        
#> 14 r-0.7.8                        
#> 15 r-0.7.9                        
#> 16 rc_0.8.0                       

# get details of open issues
get_gh_issues("dplyr", "tidyverse")

#> # A tibble: 30 x 4
#>    date       number author      title                                          
#>    <date>      <int> <chr>       <chr>                                          
#>  1 2019-07-03   4460 DavisVaugh~ False positive implicit NA warning with factor~
#>  2 2019-07-02   4459 bschneidr   No longer able to use case_when() in rename_al~
#>  3 2019-07-01   4457 dhicks      Gender sex                                     
#>  4 2019-07-01   4456 dhicks      "Confusing/offensive use of \"gender\" in `sta~
#>  5 2019-07-01   4455 dhicks      null vehicles in `starwars` dataset constructi~
#>  6 2019-06-29   4453 krlmlr      WIP: Fix character encoding problems           
#>  7 2019-06-29   4452 krlmlr      Review documentation for naming in scoped func~
#>  8 2019-06-29   4451 ajkroeg     modify() within group_by() in 0.8.2 not workin~
#>  9 2019-06-29   4450 krlmlr      Move responsibility of an eval_tidy() call     
#> 10 2019-06-27   4446 rcorty      bind_rows drops names of list columns          
#> # ... with 20 more rows

# get labels
get_gh_labels("dplyr", "tidyverse")

#> # A tibble: 16 x 2
#>    name                                   color 
#>    <chr>                                  <chr> 
#>  1 architecture                           eeeeee
#>  2 breaking change :skull_and_crossbones: E0B3A2
#>  3 bug                                    E0B3A2
#>  4 data frame                             eeeeee
#>  5 documentation                          CBBAB8
#>  6 feature                                B4C3AE
#>  7 generic                                eeeeee
#>  8 good first issue :heart:               CBBAB8
#>  9 help wanted :heart:                    C5C295
#> 10 hybrid                                 eeeeee
#> 11 join :twisted_rightwards_arrows:       eeeeee
#> 12 performance :rocket:                   eeeeee
#> 13 reprex                                 C5C295
#> 14 tidy-dev-day :nerd_face:               CBBAB8
#> 15 vector                                 eeeeee
#> 16 wip                                    E1B996

# get milestones
get_gh_milestones("dplyr", "tidyverse")

#> # A tibble: 4 x 6
#>   title  tart_date  due_date description               open_issues closed_issues
#>   <chr>  <date>     <lgl>    <chr>                           <int>         <int>
#> 1 blues~ 2013-11-20 NA       Things that would be awe~           0            16
#> 2 future 2015-10-21 NA       ""                                  2            51
#> 3 0.9.0  2018-06-01 NA       ""                                 31            10
#> 4 0.8.3  2019-06-25 NA       ""                                  1             2

# get code of conduct
get_gh_coc("dplyr", "tidyverse")

#> # Contributor Code of Conduct
#> 
#> As contributors and maintainers of this project, we pledge to respect all people who 
#> contribute through reporting issues, posting feature requests, updating documentation,
#> submitting pull requests or patches, and other activities.
#> 
#> We are committed to making participation in this project a harassment-free experience for
#> everyone, regardless of level of experience, gender, gender identity and expression,
#> sexual orientation, disability, personal appearance, body size, race, ethnicity, age, or religion.
#> 
#> Examples of unacceptable behavior by participants include the use of sexual language or
#> imagery, derogatory comments or personal attacks, trolling, public or private harassment,
#> insults, or other unprofessional conduct.
#> 
#> Project maintainers have the right and responsibility to remove, edit, or reject comments,
#> commits, code, wiki edits, issues, and other contributions that are not aligned to this 
#> Code of Conduct. Project maintainers who do not follow the Code of Conduct may be removed 
#> from the project team.
#> 
#> Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by 
#> opening an issue or contacting one or more of the project maintainers.
#> 
#> This Code of Conduct is adapted from the Contributor Covenant 
#> (http://contributor-covenant.org), version 1.0.0, available at 
#> http://contributor-covenant.org/version/1/0/0/

# get license
get_gh_license("dplyr", "tidyverse")

# get details of open pull requests
get_gh_pr("dplyr", "tidyverse")

#> # A tibble: 10 x 4
#>    number date       title                                                status
#>     <int> <date>     <chr>                                                <chr> 
#>  1   4457 2019-07-01 Gender sex                                           open  
#>  2   4453 2019-06-29 WIP: Fix character encoding problems                 open  
#>  3   4452 2019-06-29 Review documentation for naming in scoped functions  open  
#>  4   4450 2019-06-29 Move responsibility of an eval_tidy() call           open  
#>  5   4416 2019-06-08 filter() gives more information on size mismatch er~ open  
#>  6   4405 2019-06-05 Update progress.R                                    open  
#>  7   4391 2019-05-28 "bind_cols() uses as_tibble(.name_repair=\"unique\"~ open  
#>  8   4136 2019-01-29 $<-.grouped_df protects grouping variables           open  
#>  9   4102 2019-01-11 pull() can also return named vectors                 open  
#> 10   4024 2018-12-17 WIP: vignette that loads tibble, reproducing bug #3~ open  

# get releases
get_gh_releases("dplyr", "tidyverse")

#> # A tibble: 24 x 4
#>    tag      date       title         prerelease
#>    <chr>    <date>     <chr>         <lgl>     
#>  1 v0.8.2   2019-07-01 dplyr 0.8.2   FALSE     
#>  2 v0.8.1   2019-05-17 dplyr 0.8.1   FALSE     
#>  3 v0.8.0.1 2019-02-15 dplyr 0.8.0.1 FALSE     
#>  4 v0.8.0   2019-02-15 dplyr 0.8.0   FALSE     
#>  5 v0.7.6   2018-10-26 dplyr 0.7.6   FALSE     
#>  6 v0.7.5   2018-05-22 dplyr 0.7.5   FALSE     
#>  7 v0.7.4   2017-09-29 dplyr 0.7.4   FALSE     
#>  8 v0.7.3   2017-09-29 dplyr 0.7.3   FALSE     
#>  9 v0.7.2   2017-07-28 dplyr 0.7.2   FALSE     
#> 10 v0.7.1   2017-06-22 dplyr 0.7.1   FALSE     
#> # ... with 14 more rows