CRAN package details

Extracts and formats package details such as title, description, version, R version dependency, imports, suggests, published date, license, authors, maintainer and associated urls from crandb API.

get_pkg_title(pkg_details)

get_pkg_desc(pkg_details)

get_pkg_version(pkg_details)

get_pkg_r_dep(pkg_details)

get_pkg_imports(pkg_details)

get_pkg_suggests(pkg_details)

get_pkg_publish_date(pkg_details)

get_pkg_license(pkg_details)

get_pkg_authors(pkg_details)

get_pkg_maintainer(pkg_details)

get_pkg_urls(pkg_details)

Arguments

pkg_details
An object of class pkg_details.

Examples

# retrieve package details from crandb api
pkgdetails <- get_pkg_details("dplyr")

#> No encoding supplied: defaulting to UTF-8.

# package title
get_pkg_title(pkgdetails)

#> [1] "A Grammar of Data Manipulation"

# package description
get_pkg_desc(pkgdetails)

#> [1] "A fast, consistent tool for working with data frame like objects,\nboth in memory and out of memory."

# package version
get_pkg_version(pkgdetails)

#> [1] "0.8.2"

# R dependency
get_pkg_r_dep(pkgdetails)

#> [1] ">= 3.2.0"

# packages imported
get_pkg_imports(pkgdetails)

#>  [1] "assertthat" "glue"       "magrittr"   "methods"    "pkgconfig" 
#>  [6] "R6"         "Rcpp"       "rlang"      "tibble"     "tidyselect"
#> [11] "utils"     

# packages suggested
get_pkg_suggests(pkgdetails)

#>  [1] "bit64"          "callr"          "covr"           "crayon"        
#>  [5] "DBI"            "dbplyr"         "dtplyr"         "ggplot2"       
#>  [9] "hms"            "knitr"          "Lahman"         "lubridate"     
#> [13] "MASS"           "mgcv"           "microbenchmark" "nycflights13"  
#> [17] "rmarkdown"      "RMySQL"         "RPostgreSQL"    "RSQLite"       
#> [21] "testthat"       "withr"          "broom"          "purrr"         
#> [25] "readr"         

# latest CRAN publish date
get_pkg_publish_date(pkgdetails)

#> [1] "2019-06-29"

# package license
get_pkg_license(pkgdetails)

#> [1] "MIT + file LICENSE"

# package authors
get_pkg_authors(pkgdetails)

#> # A tibble: 5 x 2
#>   author          role    
#>   <chr>           <chr>   
#> 1 Hadley Wickham  aut, cre
#> 2 Romain François aut     
#> 3 Lionel Henry    aut     
#> 4 Kirill Müller   aut     
#> 5 RStudio         cph, fnd

# package maintainer
get_pkg_maintainer(pkgdetails)

#> # A tibble: 1 x 2
#>   name              email             
#>   <chr>             <chr>             
#> 1 "Hadley Wickham " hadley@rstudio.com

# urls associated with the package
get_pkg_urls(pkgdetails)

#> # A tibble: 3 x 2
#>   website urls                                     
#>   <chr>   <chr>                                    
#> 1 Bugs    https://github.com/tidyverse/dplyr/issues
#> 2 Others  http://dplyr.tidyverse.org               
#> 3 GitHub  https://github.com/tidyverse/dplyr